fix: StudyGroup ID 타입 변경에 따른 수정
- StudyGroupQueryRepository에서 findDetailById 메서드의 파라미터 타입을 Integer에서 Long으로 변경
- StudyGroupService에서 해당 메서드를 호출하는 부분 수정

+++ src/main/java/com/depth/learningcrew/domain/studygroup/repository/StudyGroupQueryRepository.java
@@ -231,7 +231,7 @@ public class StudyGroupQueryRepository {
         return paginateByType(searchConditions, user, pageable, StudyGroupFilterType.ALL);
     }
 
-    public Optional<StudyGroup> findDetailById(Integer groupId) {
+    public Optional<StudyGroup> findDetailById(Long groupId) {
         return Optional.ofNullable(queryFactory.selectFrom(studyGroup)
                 .leftJoin(studyGroup.owner, user).fetchJoin()
                 .leftJoin(studyGroup.categories, groupCategory).fetchJoin()
